chore: reorganize baggage folder
